refactor: 80%
