+++ components/table/style/index.less
@@ -186,4 +186,7 @@
         cursor: pointer;
         transition: all @animation-duration-base @ease-in-out;
     }
+    .fes-scrollbar-track {
+        z-index: 100;
+    }
 }
