+++ docs/.vitepress/components/table/index.md
@@ -10,7 +10,7 @@ app.use(FTable);
 
 ## 代码演示
 
-<!-- ### 基础用法
+### 基础用法
 
 当 `f-table` 元素中注入 `data` 对象数组后，在 `f-table-column` 中用 `prop` 属性来对应对象中的键名即可填入数据，用 `label` 属性来定义表格的列名。可已使用`width` 属性来定义列宽。
 
@@ -63,7 +63,7 @@ app.use(FTable);
 
 自定义列的显示内容，可组合其他组件使用。
 
---TEMPLATE -->
+--TEMPLATE
 
 ### 虚拟滚动
